+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename `IdsGenerator` to `IdGenerator`
   ([#1651])(https://github.com/open-telemetry/opentelemetry-python/pull/1651)
+- Make TracerProvider's resource attribute private
+  ([#1652])(https://github.com/open-telemetry/opentelemetry-python/pull/1652)
 - Rename Resource's `create_empty` to `get_empty`
   ([#1653])(https://github.com/open-telemetry/opentelemetry-python/pull/1653)